backend/pyproject.toml
[project]
name="backend"
version="0.1.0"
description=""
authors=[
    {name = "Everaldo Junior"}
]
readme = "README.md"
requires-python=">=3.12.3"
dynamic=["dependencies"]

[tool.poetry.dependencies]
django = "^5.2"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
google-api-python-client = "^2.169.0"
google-auth = "^2.39.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
requests = "^2.32.3"
matplotlib = "^3.10.1"
numpy = "^2.2.5"
mysqlclient = "^2.2.7"
drf-yasg = "^1.21.10"
python-decouple = "^3.8"
django-filter = "^25.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.6"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.11.1"
selenium = "^4.32.0"
pytest-playwright = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
